chg: move last "j2x_creation" test to similar j2x test file
- there does not seem to be a meaningful difference between the purpose
  and structure of this test and those in j2x_instantiation
